refactor: personaldata routes to reflect simpler module structure
