--- web-client/src/modules/personalData/pages/routes/RoleInfoRoute/constants.tsx
@@ -1,3 +1,0 @@
-import Location from 'react-app-location';
-
-export const RoleInfoLocation = new Location('/personal-data/role-info');
