+++ web-client/src/pages/routes/ProtectedRoute.tsx
@@ -4,8 +4,10 @@ import { Redirect, Route, RouteProps, useLocation } from 'react-router-dom';
 import { observeUserAction } from 'src/ducks/auth/actions';
 import { observePrivileged, observeProfile } from 'src/ducks/profile/actions';
 import { ProfileState } from 'src/ducks/profile/types';
-import { PersonalDataLocation } from 'src/modules/personalData/pages/routes/PersonalDataRoute/constants';
-import { RoleInfoLocation } from 'src/modules/personalData/pages/routes/RoleInfoRoute/constants';
+import {
+  PersonalDataLocation,
+  RoleInfoLocation,
+} from 'src/modules/personalData/constants';
 import { AppState } from 'src/store';
 
 import LoadingWrapper from '../../components/LoadingComponent/LoadingComponent';
